+++ src/index.ts
@@ -47,48 +47,60 @@ function main(): void {
 }
 
 function initMaidr(maidr: Maidr, plot: HTMLElement): void {
-  let maidrRoot: HTMLElement | null = null;
+  let maidrContainer: HTMLElement | null = null;
+  let reactContainer: HTMLElement | null = null;
   let controller: Controller | null = null;
 
-  const onBlur = (event: FocusEvent): void => {
-    if (!controller || !controller.shouldDispose(event)) {
-      return;
-    }
+  const onFocusOut = (): void => {
+    // Calling `setTimeout` to let React's rendering pipeline
+    // to process all the events to finalize the correct activeElement.
+    setTimeout(() => {
+      if (!maidrContainer) {
+        return;
+      }
 
-    controller?.dispose();
-    controller = null;
+      const activeElement = document.activeElement as HTMLElement;
+      if (!maidrContainer.contains(activeElement)) {
+        controller?.dispose();
+        controller = null;
+      }
+    }, 0);
   };
-  const onFocus = (): void => {
-    if (!maidrRoot) {
+  const onFocusIn = (): void => {
+    if (!maidrContainer || !reactContainer) {
       return;
     }
 
     if (!controller) {
       // Create a deep copy to prevent mutations on the original maidr object.
       const maidrClone = JSON.parse(JSON.stringify(maidr));
-      controller = new Controller(maidrClone, maidrRoot, plot);
+      controller = new Controller(maidrClone, plot, reactContainer);
     }
   };
 
   const figureElement = document.createElement(Constant.FIGURE);
-  figureElement.id = Constant.MAIDR_FIGURE + maidr.id;
+  figureElement.id = `${Constant.MAIDR_FIGURE}-${maidr.id}`;
   plot.parentNode!.replaceChild(figureElement, plot);
   figureElement.appendChild(plot);
 
   const articleElement = document.createElement(Constant.ARTICLE);
-  articleElement.id = Constant.MAIDR_ARTICLE + maidr.id;
+  articleElement.id = `${Constant.MAIDR_ARTICLE}-${maidr.id}`;
   figureElement.parentNode!.replaceChild(articleElement, figureElement);
   articleElement.appendChild(figureElement);
 
-  maidrRoot = figureElement;
-  plot.addEventListener(DomEventType.FOCUS_IN, onFocus);
-  plot.addEventListener(DomEventType.CLICK, onFocus);
-  plot.addEventListener(DomEventType.FOCUS_OUT, onBlur);
+  reactContainer = document.createElement(Constant.DIV);
+  reactContainer.id = `${Constant.REACT_CONTAINER}-${maidr.id}`;
+  figureElement.appendChild(reactContainer);
+
+  maidrContainer = figureElement;
+  plot.addEventListener(DomEventType.FOCUS_IN, onFocusIn);
+  plot.addEventListener(DomEventType.CLICK, onFocusIn);
+  maidrContainer.addEventListener(DomEventType.FOCUS_OUT, onFocusOut);
 
   (() => {
     // Create a deep copy to prevent mutations on the original maidr object.
     const maidrClone = JSON.parse(JSON.stringify(maidr));
-    const controller = new Controller(maidrClone, maidrRoot, plot);
+    const controller = new Controller(maidrClone, plot, reactContainer);
     controller.dispose();
   })();
 }