+++ src/state/viewModel/registry.ts
@@ -1,3 +1,6 @@
+import type { BrailleViewModel } from '@state/viewModel/brailleViewModel';
+import type { DisplayViewModel } from '@state/viewModel/displayViewModel';
+import type { ReviewViewModel } from '@state/viewModel/reviewViewModel';
 import type { Disposable } from '@type/disposable';
 import type { ChatViewModel } from './chatViewModel';
 import type { HelpViewModel } from './helpViewModel';
@@ -5,8 +8,11 @@ import type { SettingsViewModel } from './settingsViewModel';
 import type { TextViewModel } from './textViewModel';
 
 export interface ViewModelMap {
+  braille: BrailleViewModel;
   chat: ChatViewModel;
+  display: DisplayViewModel;
   help: HelpViewModel;
+  review: ReviewViewModel;
   settings: SettingsViewModel;
   text: TextViewModel;
 }