+++ src/ui/pages/Text.tsx
@@ -1,14 +1,14 @@
-import type { FC } from 'react';
 import { useViewModelState } from '@state/hook/useViewModel';
+import { Constant } from '@util/constant';
 import React from 'react';
 
-const Text: FC = () => {
+const Text: React.FC = () => {
   const { enabled, announce, value, message } = useViewModelState('text');
   const shouldAnnounce = announce || message;
 
   return (
     <div
-      id="maidr-text-container"
+      id={Constant.TEXT_CONTAINER}
       {...(shouldAnnounce && {
         role: 'alert',
       })}